requirements.txt
numpy
scipy
faiss-cpu
